Devide trial tests
